requirements.txt
torch==1.8.0
basicsr==1.4.2
realesrgan==0.3.0
opencv-python==4.5.1
ffmpeg-python==0.2.0
rich==10.0.0